package/pkg-cargo: fix build with cc-rs
cc-rs (https://docs.rs/cc/latest/cc/) is widely used in Rust crates to
compile non-Rust files and handle cross-compilation. However such crate
could be either included in the target application and/or in build
scripts (build.rs). When the latter happens, $CC and/or $CLFAGS, if set,
will be used even though the target is the host. Fix those cases by
providing an host specific CC and CFLAGS.

Signed-off-by: Antoine Tenart <atenart@kernel.org>